gprofng: add the gprofng tool
This commit adds a new tool to binutils: the gprofng profiler.  It
supports profiling C, C++ and Java programs.  Unlike the old gprof, it
doesn't require to build annotated versions of the programs:  profiling
"production" binaries works just fine.

The tool is composed by a couple of libraries, a set of utilities named
gp-* and a driver program called `gprofng'.

Tests and documentation are added in subsequent commits.

gprofng/ChangeLog
2021-08-08  Vladimir Mezentsev  <vladimir.mezentsev@oracle.com>

	* Makefile.def: Add the gprofng build.
	* configure.ac: Add gprofng in host_tools.
	* Makefile.in: Rebuild.
	* configure: Rebuild.
	* binutils/MAINTAINERS: Add the gprofng maintainer.
	* gprofng: New directory.