gprofng: add the gprofng tool
This commit adds a new tool to binutils: the gprofng profiler.  It
supports profiling C, C++ and Java programs.  Unlike the old gprof, it
doesn't require to build annotated versions of the programs:  profiling
"production" binaries works just fine.

The tool is composed by a couple of libraries, a set of utilities named
gp-* and a driver program called `gprofng'.

Tests and documentation are added in subsequent commits.

gprofng/ChangeLog
2022-02-02  Vladimir Mezentsev  <vladimir.mezentsev@oracle.com>

	* gprofng: New directory.
	* include/collectorAPI.h: New file. Interface to libgp-collectorAPI.
	* include/libcollector.h: New file. Interface to libgp-collector.
	* include/libfcollector.h: Likewise.